+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Parsing multi-file programs is now parallelised, making it
   *slightly* faster.
 
+* Reloading a large program in `futhark repl` is now faster, as long
+  as not too many of its files have been modified (#1597).
+
 ### Removed
 
 ### Changed